requirements.txt
requests
PyGithub
google-generativeai